Minor tweaks. Plan deceleration in advance
There was a misperfection:
Since we do not want to decelerate most of the time, I prohibited
deceleration moves in deep calculation. But sometimes we do need
break. In order to do so, I permitted deceleration move on the
first step.

It appeared that since on a computed plan we cannot decelerate
upon approaching the target, we have to decelerate in advance(

This way we reach the target later and losing points.

To fix this, I had to allow deceleration moves in future steps.